requirements.txt
streamlit>=1.36
google-genai>=0.3.0
pydantic>=2.6
